Moved syntax APIs separate folder
