+++ bagel/__main__.py
@@ -4,9 +4,9 @@ import colorama
 from colorama import Fore
 
 from .codeanalysis.evaluator import Evaluator
-from .codeanalysis.syntaxnode import SyntaxNode
-from .codeanalysis.syntaxtoken import SyntaxToken
-from .codeanalysis.syntaxtree import SyntaxTree
+from .codeanalysis.syntax.syntaxnode import SyntaxNode
+from .codeanalysis.syntax.syntaxtoken import SyntaxToken
+from .codeanalysis.syntax.syntaxtree import SyntaxTree
 
 colorama.init(autoreset=True)
 show_tree = False
