+++ bagel/codeanalysis/syntax/lexer.py
