+++ bagel/codeanalysis/syntax/syntaxtoken.py
