+++ bagel/codeanalysis/syntax/syntaxtree.py
